UserBackend/auth_middleware.py
import logging
from functools import wraps
import jwt
from flask import request, abort
from flask import current_app
from model.User import Userv2Class

def token_required(f):
    @wraps(f)
    def decorated(*args, **kwargs):
        token = None
        logging.info(request.headers)
        if "Authorization" in request.headers:
            token = request.headers["Authorization"].split(" ")[1]
        if not token:
            return {
                "message": "Authentication Token is missing!",
                "data": None,
                "error": "Unauthorized"
            }, 401
        try:
            data=jwt.decode(token, current_app.config["SECRET_KEY"], algorithms=["HS256"])
            current_user=Userv2Class().get_user_by_id(data["_id"])
            if current_user is None:
                return {
                "message": "Invalid Authentication token!",
                "data": None,
                "error": "Unauthorized"
            }, 401
            if not current_user["active"]:
                abort(403)
        except Exception as e:
            logging.exception(e)
            return {
                "message": "Something went wrong",
                "data": None,
                "error": str(e)
            }, 500

        return f(current_user, *args, **kwargs)

    return decorated