UserBackend/requirements.txt
flask
PyJWT
Flask-Compress
flask-cors
pymongo